+++ Graph-Depth_Breath_First_Search/Graph.h
@@ -2,7 +2,7 @@
 #define GRAPH_H
 
 
-
+#include <stdint.h>
 #include "DLinkedList.h"
 
 enum {A, B, C, D, E, F, G, H, I, J};
@@ -15,7 +15,7 @@ typedef struct _ual
     int* visitInfo;
 }ALGraph;
 
-void GraphInit(ALGraph *pg, int nv);
+void GraphInit(ALGraph *pg, uint8_t nv);
 void GraphDestory(ALGraph *pg);
 void AddEdge(ALGraph *pg, int fromV, int toV);
 void ShowGraphEdgeInfo(ALGraph * pg);
